% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharityNavigatoR-package.R
\docType{package}
\name{CharityNavigatoR-package}
\alias{CharityNavigatoR}
\alias{CharityNavigatoR-package}
\title{CharityNavigatoR: Charity Navigator API Client}
\description{
Founded in 2001, Charity Navigator has become the nation's largest and most-utilized evaluator of charities.
        The company provides charity data and a proprietery rating system through a REST API.
        The CharityClient package provides an family of R functions that access the API service, making it accessible
        and R-friendly for R researchers.
        See https://charity.3scale.net/docs/data-api/reference#api-documentation for the Charity Navigator API schema.
}
\author{
\strong{Maintainer}: Jose Andres Montes Lopez \email{jam2448@columbia.edu}

}
\keyword{internal}
