% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIRS.R
\name{GetIRS}
\alias{GetIRS}
\title{Access Charity IRS Data}
\usage{
GetIRS(name = NULL, state = NULL, sizeRange = NULL, limit = 10)
}
\arguments{
\item{name}{search for a near match based on string.}

\item{state}{A two letter name.}

\item{sizeRange}{Search organizations within a bracket,
as measured in annual total expenses.
1 = Up to $3.5MM,
2 = $3.5MM to $13.5MM,
3 = $13.5MM and up}

\item{limit}{Number of organizations that will be listed.}
}
\description{
This function allows the users to access IRS charity data from Charity Navigator API
through a variety of different parameters. Parameters should be left null if it is not of interest
so as to include all fields that have that parameter.
}
\details{
Returns a data.table with 21 columns filtered based on search query.
All available IRS data will be included.
}
\examples{
GetIRS()

GetIRS(state = "CA")

GetIRS(name = "Refresh", state = "CA")
}
