% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOrganizations.R
\name{GetOrganizations}
\alias{GetOrganizations}
\title{Access All Available Charity Data}
\usage{
GetOrganizations(name = NULL, state = NULL, sizeRange = NULL,
  limit = 10)
}
\arguments{
\item{name}{search for a near match based on string.}

\item{state}{A two letter name.}

\item{sizeRange}{Search organizations within a bracket,
as measured in annual total expenses.
1 = Up to $3.5MM,
2 = $3.5MM to $13.5MM,
3 = $13.5MM and up}

\item{limit}{Number of organizations that will be listed.}
}
\description{
This function allows the users to access charity data from Charity Navigator API
through a variety of different parameters. Parameters should be left null if it is not of interest
so as to include all fields that have that parameter.
}
\details{
Returns a data.table with 39 columns filtered based on search query.
Certain columns will be empty unless user has a premium subscription.
}
\examples{
GetOrganizations()

GetOrganizations(state = "CA",10)

GetOrganizations(name = "Refresh")
}
